Skip .venv and .git when searching workspace
Skip .venv since it can contain stray pyprojects that are not related to
the current project. Further items could be ignored here.

Correctness fix (previously pyproject.toml installed into some package
in the environment, could ruin the sync), as well as a speed up since we
can skip scanning all the files in the environment.