workspace is member: fix relative path (#300)
It looks like the relative path is always empty. I'm not 100% sure what
the intention was here with the canonicalization.

With this fix, limiting members using an explicit list in the workspace
should work better. Previously it seemed to include everything in all
case.

rye show is the quickest way to see the workspace member list, but it
affects rye lock foremost (with bad effects if it includes stray
projects.)