add --modify-path and --no-modify-path to rye self install
This adds the option to never modify paths to rye install, even
when `--yes` is provided. This addresses #620.

We offer a --modify-path and a --no-modify-path option to either
always modify or never modify. Providing --yes, will be the same
as --modify-path. This is in line with other command line tools
like curl, git and ripgrep.

We do this by adding a new enum YesNoArg, which can be obtained by
providing an ArgGroup with respective modify_path and no_modify_path
boolean values.