utils: Add a quieter() function to CommandOutput
CommandOutput now gets a quieter() function. There are some contexts
in which a CommandOutput::Normal means that -q flag should be passed,
e.g. in sync. At the moment this is handled on a per-callsite base.
However, as we are moving to an abstracted Uv in uv.rs, we need a way
to move from a CommandOutput::Normal to Quiet in these contexts.