sync: move update_venv_sync_marker to utils
We are moving update_venv_sync_marker to utils. It's self contained
within utils and relies on `mark_path_sync_ignore`. This paves the
way for us to use it in uv.rs more easily, where we will offer a method
on UvWithVirtual that calls `update_venv_sync_marker` for the venv.