sources: Add support for environment
Linux and in rare cases windows binaries can be built with different
environments. At the moment rye assumes we will be always using a gnu
environment on linux. This is not true for distributions like Alpine,
which use musl instead of glibc.

Indygreg offers musl builds, but we don't pull them in just yet.

1. We introduce the notion of a "environment", which can either be unset (None)
   or set to a string.
2. We set the environment string to Some("gnu") for all existing linux downloads
3. We set the environment string to None for all others
4. We modified match_versions to correctly resolve environment or fallback
   to the default environment (Some("gnu") for linux, None for all others).
5. We added tests.