py: env selection
We are using the target environment to select the requested cpython
version. Since `std::env::consts` does not offer a way to select
the target os on runtime, we are having to find a suitable way.

This implies that a rye version build for musl will pull
a musl toolchain by default and a rye version for gnu will will
pull a gnu toolchain. In general this should work. On a fully
musl system such as alpine, you wouldn't have a system that is compatible
with gnu anyway and on a GNU system you would just always pull
a musl toolchain which will work.