refactor: Add fallback prompts
This refactors the section that handles prompting the user. To avoid tangled logic here, and in other parts of the code, we want to first establish if encryption or decryption should happen. This influences the prompts needed as well as give us the ability to move where these operations end up ocurring -- which cleans this up nicely. Our functions now return `None` if the user decides not to provide an input.