Use uv in rye build when enabled
The `build` package supports uv since version 1.2.0 - it can use either
uv or pip as installer.

The main way to use it would be to install `build[uv]` but rye already
has uv installed separately, build will find it on the path as a
fallback, so this can work for rye.