Bump boostrap ruff to 0.4.4 (#1075)
Noticed ruff was a bit old and was missing a lot of rules, leading to a
lot of "unused noqa".

I unfortunately dont have the rust knowledge to improve this, but it
would be nice if we could update ruff independently of rye and still run
it using `rye lint/fmt`.

Thanks for this amazing tool! Gave it a shot today and really really
enjoyed playing around with it and testing it out. Currently porting
some of my projects to use it!