Move from `rye-up.com` to `rye.astral.sh` (#1113)
## Summary

We're having issues with `rye-up.com`. We're trying to recover it but
need to work with the registrar. In the interim, I'm going to move the
docs and installers to `rye.astral.sh`. When `rye-up.com` is recovered,
we can redirect, either from `rye.astral.sh` to `rye-up.com` or vice
versa.

Closes #1111.