Add ability to specify option to generate hashes within pyproject.toml (#1129)
Please note: I haven't written any Rust before, but just wanted this
feature hence thought I'd try submitting a PR for this. I've essentially
tried to mimic the code regarding the `--with-sources` feature. I'm not
sure how to compile and test locally (so welcome any info on this - and
if helpful I would be happy to test this locally and get back to confirm
if this works).

All thoughts/help/feedback etc. very welcome, and no worries if this
feature is not useful.

- Adds ability to specify option to generate hashes via the
pyproject.toml (similar to
https://rye.astral.sh/guide/pyproject/#toolryelock-with-sources).
- Updates doc regarding the `--generate-hashes` option too.
- Added [`+++
0.35.0`](https://github.com/astral-sh/rye/pull/1129/files#diff-8ea7c706930a01df775defe815afe2478a41699e0ac0b99d7b069c0fe0d9f45fR65)
to the doc in case it gets merged for this version, but feel free to
amend as appropriate.

Thanks!