fix verify-ssl option parsing problem (#1208)
https://github.com/astral-sh/rye/blob/main/docs/guide/sources.md?plain=1#L131

The documentation refers to verify-ssl to avoid ssl verification, but
during parsing, it refers to verify_ssl. This PR will allow the
installation of packages from non-SSL PyPI repositories.
If the documentation needs to be corrected instead of the source code,
please let me know, and I will commit the changes.