Update Docs Description for Download Options (#1201)
Was just downloading Rye to try it out and got a bit annoyed by the
confusing download links (like `x64 Intel Windows` - I have an AMD CPU,
and it obviously doesn't matter), so I updated them to more standard
terms, although there are several options to update them to, so feel
free to ask for / make changes).

I also think it might be a good idea to consider dropping the Windows
x86 (32-bit) builds, as Windows 11 doesn't even support x86, and Windows
10 EOL is next year (plus, I doubt there are many Windows 10 users using
the 32-bit version. Especially developers), and it just confuses
newcomers with an additional option to pick from.