Retire Rye and add a uv migration guide
https://lucumr.pocoo.org/2024/8/21/harvest-season/

> Domination is a goal because it means that most investment will go
> into one stack. I can only re-iterate my wish and desire that Rye (and
> with it a lot of other tools in the space) should cease to exist once
> the dominating tool has been established. For me uv is poised to be
> that tool. It's not quite there today yet for all cases, but it will
> be in no time, and now is the moment to step up as a community and
> start to start to rally around it. That doesn't mean that this tool
> will be the tool forever. Things come and go and maybe there is a
> future for some other tool.
>
> But today I'm looking forward to the moment when there will be a final
> release of Rye that is no remaining functionality other than to just
> largely alias to uv, that retires Rye specific functionality and
> migrates you over to uv.

This does not do a final release - adding a compatibility layer for uv
and an automatic migration is likely to be buggy and incomplete, and
users are better served by changing the command they type to `uv`.
Instead, it adds a migration guide based on the discussion thread
(https://github.com/astral-sh/rye/discussions/1342), documenting things
that are not immediately obvious to a user swapping out `rye` for `uv`.